refactor error handling.

+++ src/controllers/carts.controller.js
@@ -7,25 +7,33 @@ import { cartsService } from "../services/carts.service.js";
 export const getCartById = async (req, res) => {
   try {
     const { cid } = req.params;
+
+    if (!cid) {
+      return res.status(400).send({
+        status: "error",
+        error: "Incomplete values",
+      });
+    }
+
     const filteredCart = await cartsService.getCartById(cid);
+
+    if (!filteredCart) {
+      return res.status(404).send({
+        status: "error",
+        error: "Failed to get to cart",
+      });
+    }
+
     return res.status(200).send({
       status: "success",
       payload: filteredCart,
     });
   } catch (error) {
-    if (
-      error.message === "Cart not found" ||
-      error.message === "Incomplete values"
-    ) {
-      return res.status(404).send({
-        status: "error",
-        error: `${error.message}`,
-      });
-    }
     console.log(`Cannot get cart with mongoose ${error}`);
-    return res
-      .status(500)
-      .send({ status: "error", error: "Failed to get cart" });
+    return res.status(500).send({
+      status: "error",
+      error: "Failed to get cart",
+    });
   }
 };
 
@@ -36,16 +44,17 @@ export const getCartById = async (req, res) => {
 export const createCart = async (req, res) => {
   try {
     const newCart = await cartsService.createCart();
+
+    if (!newCart) {
+      return res.status(404).send({
+        status: "error",
+        error: "Failed to create cart",
+      });
+    }
+
     res.status(201).send({ status: "Success", payload: newCart });
   } catch (error) {
     console.log(`Failed to create cart with mongoose ${error}`);
-
-    if (error.message === "Failed to create cart") {
-      return res.status(404).send({
-        status: "error",
-        message: { error: "Failed to create cart" },
-      });
-    }
     return res
       .status(500)
       .send({ status: "error", error: "Failed to create cart" });
@@ -56,21 +65,28 @@ export const addToCart = async (req, res) => {
   try {
     const { cid, pid } = req.params;
     const { quantity } = req.body;
+
+    if (!cid || !pid) {
+      return res.status(400).send({
+        status: "error",
+        error: "Incomplete values",
+      });
+    }
+
     const productAddedToCart = await cartsService.addToCart(cid, pid, quantity);
+
+    if (!productAddedToCart) {
+      return res.status(404).send({
+        status: "error",
+        error: "Failed to add product to cart",
+      });
+    }
+
     return res
       .status(201)
       .send({ status: "Success", payload: productAddedToCart });
   } catch (error) {
     console.log(`Cannot add to cart with mongoose ${error}`);
-    if (
-      error.message === "Incomplete values" ||
-      error.message === "Failed to add product to cart"
-    ) {
-      return res.status(404).send({
-        status: "error",
-        error: `${error.message}`,
-      });
-    }
     return res
       .status(500)
       .send({ status: "error", error: "Failed to add product to cart" });
@@ -86,7 +102,21 @@ export const updateCart = async (req, res) => {
     const { cid } = req.params;
     const products = req.body;
 
+    if (!cid || !products) {
+      return res.status(400).send({
+        status: "error",
+        error: "Incomplete values",
+      });
+    }
+
     const updatedCart = await cartsService.updateCart(cid, products);
+
+    if (!updatedCart) {
+      return res.status(404).send({
+        status: "error",
+        error: "Failed to update cart",
+      });
+    }
 
     return res.status(200).send({
       status: "success",
@@ -94,23 +124,6 @@ export const updateCart = async (req, res) => {
     });
   } catch (error) {
     console.log(`Cannot update cart with mongoose ${error}`);
-
-    if (error.message === "Incomplete values") {
-      return res.status(400).send({
-        status: "error",
-        message: { error: "Incomplete values" },
-      });
-    }
-
-    if (error.message === "Could not update cart. No cart found") {
-      return res.status(404).send({
-        status: "error",
-        message: {
-          error: "Could not update cart. No cart found",
-        },
-      });
-    }
-
     return res
       .status(500)
       .send({ status: "error", error: "Failed to update cart" });
@@ -122,11 +135,25 @@ export const updateProductFromCart = async (req, res) => {
     const { cid, pid } = req.params;
     const { quantity = 1 } = req.body;
 
-    let updatedProductFromCart = await cartsService.updateProductFromCart(
+    if (!cid || !pid) {
+      return res.status(400).send({
+        status: "error",
+        error: "Incomplete values",
+      });
+    }
+
+    const updatedProductFromCart = await cartsService.updateProductFromCart(
       cid,
       pid,
       quantity
     );
+
+    if (updatedProductFromCart.modifiedCount === 0) {
+      return res.status(404).send({
+        status: "error",
+        error: "Product was not found in that cart",
+      });
+    }
 
     return res.status(200).send({
       status: "success",
@@ -134,22 +161,6 @@ export const updateProductFromCart = async (req, res) => {
     });
   } catch (error) {
     console.log(`Cannot update cart with mongoose ${error}`);
-    if (error.message === "Incomplete values") {
-      return res.status(400).send({
-        status: "error",
-        message: { error: "Incomplete values" },
-      });
-    }
-
-    if (error.message === "Product was not found in that cart") {
-      return res.status(404).send({
-        status: "error",
-        message: {
-          error: "Product was not found in that cart",
-        },
-      });
-    }
-
     return res
       .status(500)
       .send({ status: "error", error: "Failed to update cart" });
@@ -164,7 +175,21 @@ export const deleteCart = async (req, res) => {
   try {
     const { cid } = req.params;
 
+    if (!cid) {
+      return res.status(400).send({
+        status: "error",
+        error: "Incomplete values",
+      });
+    }
+
     const deletedCart = await cartsService.deleteCart(cid);
+
+    if (!deletedCart) {
+      return res.status(404).send({
+        status: "error",
+        error: "Cart not found",
+      });
+    }
 
     return res.status(200).send({
       status: "success",
@@ -172,23 +197,6 @@ export const deleteCart = async (req, res) => {
     });
   } catch (error) {
     console.log(`Cannot delete cart with mongoose ${error}`);
-
-    if (error.message === "Incomplete values") {
-      return res.status(400).send({
-        status: "error",
-        message: { error: "Incomplete values" },
-      });
-    }
-
-    if (error.message === "Cart not found") {
-      return res.status(404).send({
-        status: "error",
-        message: {
-          error: "Cart not found",
-        },
-      });
-    }
-
     return res
       .status(500)
       .send({ status: "error", error: "Failed to delete cart" });
@@ -199,10 +207,24 @@ export const deleteProductFromCart = async (req, res) => {
   try {
     const { cid, pid } = req.params;
 
+    if (!cid || !pid) {
+      return res.status(400).send({
+        status: "error",
+        error: "Incomplete values",
+      });
+    }
+
     const deletedProductFromCart = await cartsService.deleteProductFromCart(
       cid,
       pid
     );
+
+    if (deletedProductFromCart.deletedCount === 0) {
+      return res.status(404).send({
+        status: "error",
+        error: "Product was not found in that cart",
+      });
+    }
 
     return res.status(200).send({
       status: "success",
@@ -210,23 +232,6 @@ export const deleteProductFromCart = async (req, res) => {
     });
   } catch (error) {
     console.log(`Cannot delete cart with mongoose ${error}`);
-
-    if (error.message === "Incomplete values") {
-      return res.status(400).send({
-        status: "error",
-        message: { error: "Incomplete values" },
-      });
-    }
-
-    if (error.message === "Product was not found in that cart") {
-      return res.status(404).send({
-        status: "error",
-        message: {
-          error: "Product was not found in that cart",
-        },
-      });
-    }
-
     return res
       .status(500)
       .send({ status: "error", error: "Failed to delete product from cart" });
